Added dummy test to check compilation and moved loaders to avoid cyclic dep
