+++ src/environment/environment_test.go
@@ -1,0 +1,7 @@
+package environment
+
+import "testing"
+
+func Test(t *testing.T) {
+	// Just to compile this module
+}
